+++ alphatwirl/misc/removal.py
@@ -25,8 +25,11 @@ def _removed(msg=''):
     def _decorate_func(f, msg):
         module_name = f.__module__
         logger = logging.getLogger(module_name)
-        name = f.__name__
-        name = '{}.{}'.format(module_name, name)
+        try: # python 3
+            name = f.__qualname__
+            name = '{}.{}'.format(module_name, name)
+        except AttributeError: # python 2
+            name = f.__name__
         text = '{}() is removed.'.format(name)
         if msg:
             text += ' ' + msg